requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
commonmark==0.9.1
cryptography==43.0.3
Deprecated==1.2.14
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.0
MarkupSafe==3.0.2
openai==1.54.5
packaging==24.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==2.5.0
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.9.0
PyNaCl==1.5.0
pyodide-cli==0.2.4
pyodide-pack==0.2.0
python-dotenv==1.0.1
requests==2.32.3
rich==12.3.0
setuptools==75.5.0
shellingham==1.5.4
sniffio==1.3.1
tqdm==4.67.0
typer==0.4.1
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
